requirements.txt
# core libraries for retrieval and generation
transformers>=4.0.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0  # or faiss-gpu if you have CUDA
accelerate>=0.15.0

# Hugging Face Hub for model downloading
huggingface_hub>=0.20.0

# SSL certificate fix (especially needed on macOS)
certifi

# optional utilities
tqdm
python-dotenv
pytest
certifi # For SSL certificate handling, especially on macOS